requirements.txt
pydantic~=2.8.2
numpy~=1.26.4
polars~=1.4.1
scikit-learn~=1.5.1
arrow~=1.3.0
pyarrow~=17.0.0
copulas~=0.11.0
fitter~=1.7.1
sdv~=1.15.0
matplotlib~=3.9.0
numpy~=1.26.4
tqdm~=4.66.4
seaborn
pandas~=2.2.2
ipykernel
kaleido
torch~=2.4.0
scipy~=1.14.0
PyYAML~=6.0.2